ENS registration stuff moved to ops directory (#435)
* move ENS registration stuff to ops directory

* documentation update
