+++ .gitignore
@@ -21,6 +21,7 @@
 /universal-login-example/src/.DS_Store
 /universal-login-example/src/views/.DS_Store
 /universal-login-example/yarn-error.log
+/universal-login-ops/.env
 /universal-login-ops/dist
 /universal-login-ops/node_modules
 /universal-login-ops/yarn-error.log
