+++ universal-login-ops/src/ENS/ENSRegistrarBase.ts
@@ -1,5 +1,5 @@
-import ENS from '@universal-login/contracts/build/ENS.json';
 import {providers, Contract, Wallet, utils, ContractFunction} from 'ethers';
+import ENS from '@universal-login/contracts/build/ENS.json';
 import {Config} from '@universal-login/commons';
 
 class ENSRegistrarBase {
@@ -8,8 +8,9 @@ class ENSRegistrarBase {
   protected variables : Record<string, string>;
   protected registrarAddress: ContractFunction | any;
   protected resolverAddress: ContractFunction | any;
+  protected provider: providers.Provider;
 
-  constructor(protected config : Config , protected readonly provider : providers.Provider, protected log = console.log) {
+  constructor(protected config: Config, provider?: providers.Provider, protected log: any = console.log) {
     this.provider = provider || new providers.JsonRpcProvider(config.jsonRpcUrl, config.chainSpec);
     this.deployer = new Wallet(config.privateKey as string, this.provider);
     this.ens = new Contract(config.chainSpec.ensAddress as string, ENS.interface, this.deployer);
