+++ universal-login-ops/test/dev/DomainRegistrar.js
@@ -1,8 +1,8 @@
 import chai, {expect} from 'chai';
 import sinonChai from 'sinon-chai';
-import basicENS from '../fixtures/basicENS';
-import DomainRegistrar from '../../lib/utils/ENS/DomainRegistrar';
-import ENSNameRegistrar from '../../lib/utils/ENS/ENSNameRegistrar';
+import basicENS from '@universal-login/relayer/test/fixtures/basicENS';
+import DomainRegistrar from '../../src/ENS/DomainRegistrar';
+import ENSNameRegistrar from '../../src/ENS/ENSNameRegistrar';
 import {utils, Contract} from 'ethers';
 import PublicResolver from '@universal-login/contracts/build/PublicResolver';
 import sinon from 'sinon';
